move pet model to model directory   - refactor test files
