+++ PsychicPets.Tests/GraphQLTests.cs
@@ -1,0 +1,6 @@
+namespace DefaultNamespace;
+
+public class GraphQLTests
+{
+    
+}
