+++ PsychicPets.Tests/PsychicPets.Tests.csproj
@@ -10,15 +10,18 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\PsychicPets\PsychicPets.csproj" />
+    <PackageReference Include="GraphQL.AspNet.TestFramework" Version="1.2.5" />
+    <PackageReference Include="Moq" Version="4.20.69" />
+    <PackageReference Include="Snapshooter.NUnit" Version="0.13.0" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.6.0" />
     <PackageReference Include="NUnit" Version="3.13.3" />
     <PackageReference Include="NUnit3TestAdapter" Version="4.4.2" />
     <PackageReference Include="NUnit.Analyzers" Version="3.6.1" />
     <PackageReference Include="coverlet.collector" Version="3.2.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\PsychicPets\PsychicPets.csproj" />
   </ItemGroup>
 
 </Project>
