+++ PsychicPets/Models/Pet.cs
@@ -1,13 +1,10 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace PsychicPets.Pets;
+namespace PsychicPets.Models;
 
 public class Pet
 {
   public int Id { get; set; }
   [Required]
   public string Name { get; set; }
-} 
-
-
-
+}
