+++ PsychicPets/Pets/PetType.cs
@@ -1,4 +1,5 @@
 using GraphQL.Types;
+using PsychicPets.Models;
 
 namespace PsychicPets.Pets;
 
@@ -9,6 +10,6 @@ public class PetType : ObjectGraphType<Pet>
         Name = "Pet";
         Description = "Pet Type";
         Field(d => d.Id, nullable: false).Description("Pet Id");
-        Field(d => d.Name, nullable: true).Description("Pet Name");
+        Field(d => d.Name, nullable: false).Description("Pet Name");
     }
 }
